+++ L801_294/L801_294.hpp
@@ -1,9 +1,6 @@
 #ifndef L801_294_h
 #define L801_294_h
-long fib_r(int n)
-{
-    if (n==0) return 0;
-    if (n==1) return 1;
-    return fib_r(n-1) + fib_r(n-2);
-}
+
+long fib_r(int n);
+
 #endif
