--- L801_294/L801_294test.cpp
@@ -1,8 +1,0 @@
-#include "L801_294.hpp"
-#include <iostream>
-
-using namespace std;
-
-int main () {
-    cout<<fib_r(5)<<" and "<<fib_r(6);
-}
