+++ L802_296/L802_296.cpp
@@ -1,6 +1,13 @@
+#include "L802_296.hpp"
+#include <iostream>
+
+using namespace std;
+
 #define MAXN 45
 #define UNKNOWN -1
-long f(MAXN+1);
+
+long f[MAXN+1];
+
 long fib_c(int n)
 {
     if (f[n] == UNKNOWN)
@@ -14,4 +21,8 @@ long fib_c_driver(int n)
     f[1] = 1;
     for (int i=2; i<=n; i++) f[i]=UNKNOWN;
     return fib_c(n);
+}
+
+int main () {
+    cout<<fib_r_driver(5);
 }
