--- L802_296/L802_296test.cpp
@@ -1,8 +1,0 @@
-#include "L802_296.hpp"
-#include <iostream>
-
-using namespace std;
-
-int main () {
-    cout<<fib_r_driver(5);
-}
