+++ L803_297/L803_297.cpp
@@ -1,5 +1,11 @@
 include "L802_296.hpp"
 
+#include <iostream>
+
+using namespace std;
+
+long f[MAXN+1];
+
 long fib_dp (int n) 
 {
     long f[MAXN+1];
@@ -7,4 +13,8 @@ long fib_dp (int n)
     f[1] = 1;
     for (int i=2; i<=n; i++) f[i]=f[i-1]+f[i-2];
     return f[n];
+}
+
+int main () {
+    cout<<fib_dp(5);
 }
