+++ L803_297/L803_297.hpp
@@ -1,15 +1,6 @@
 #ifndef L803_297_h
 #define L803_297_h
 
-#include "L802_296.hpp"
-
-long fib_dp (int n) 
-{
-    long f[MAXN+1];
-    f[0] = 0;
-    f[1] = 1;
-    for (int i=2; i<=n; i++) f[i]=f[i-1]+f[i-2];
-    return f[n];
-}
+long fib_dp (int n);
 
 #endif
