--- L803_297/L803_297test.cpp
@@ -1,8 +1,0 @@
-#include "L803_297.hpp"
-#include <iostream>
-
-using namespace std;
-
-int main () {
-    cout<<fib_dp(5);
-}
