--- L804_298/L804_298test.cpp
@@ -1,8 +1,0 @@
-#include "L804_298.hpp"
-#include <iostream>
-
-using namespace std;
-
-int main () {
-    cout<<fib_ultimate(5);
-}
