--- L805_300/L805_300test.cpp
@@ -1,13 +1,0 @@
-#include "L805_300.hpp"
-#include <iostream>
-
-using namespace std;
-
-int main() {
-    int n, m;
-    cout<<"choose n"<<endl;
-    cin>>n;
-    cout<<"choose the number of coefficient"<<endl;
-    cin>>m;
-    cout<<binomial_coefficient(n,m);
-}
