--- L806_301/L806_301test.cpp
@@ -1,8 +1,0 @@
-#include "L806_301.hpp"
-#include <iostream>
-
-using namespace std;
-
-int main () {
-    /*ghtlcnfdbv? xnj pltcm rjl*/
-}
