--- L806_301/L808_303.cpp
@@ -1,31 +1,0 @@
-#include "L812_307.hpp"
-#include "L807_302.hpp"
-#include "L811_307.hpp"
-#include "L806_301.hpp"
-
-goal_cell(char *s, char *t, int *i, int *j);
-match(char *s, char *t);
-indel(char *s);
-
-
-int string_compare(char *s, char *t)
-{
-	int i, j, k;
-	int opt[3];
-	for (size_t i=0; i<strlen(s); i++) {
-		for (size_t j=0; j<strlen(t); j++) {
-			opt[MATCH] = m[i-1][j-1].cost + match(s[i],t[i]);
-			opt[INSERT] = m[i][j-1].cost + indel(t[j]);
-			opt[DELETE] = m[i-1][j].cost + indel(s[i]);
-			m[i][j].cost = opt[MATCH];
-			m[i][j].parent = MATCH;
-			for (size_t k=INSERT; k<=DELETE; k++)
-				if (opt[k] < m[i][j].cost) {
-					m[i][j].cost = opt[k];
-					m[i][j].parent = k;
-				}
-		}
-	}
-	goal_cell(s, t, &i, &j);
-	return (m[i][j].cost);
-}
