+++ L807_302/L807_302.cpp
@@ -1,3 +1,8 @@
+#include "L807_302.hpp"
+#include <iostream>
+
+using namespace std;
+
 #define MAXLEN 2 // Изменяемое значение, которое Скиена не задал в этом параграфе
 
 typedef struct {
@@ -5,5 +10,9 @@ typedef struct {
 	int parent; /* Родительская ячейка*/
 } cell;
 
-
-cell m[MAXLEN+1][MAXLEN+1]; /* Таблица динамического программирования*/
+int main () {
+	cell m[MAXLEN+1][MAXLEN+1]; /* Таблица динамического программирования*/
+    m[0][0].cost   = 1;
+    m[0][0].parent = 2;
+    cout<<m[0][0].cost<<" () "<<m[0][0].parent;
+}
