--- L807_302/L807_302test.cpp
@@ -1,10 +1,0 @@
-#include "L807_302.hpp"
-#include <iostream>
-
-using namespace std;
-
-int main () {
-    m[0][0].cost   = 1;
-    m[0][0].parent = 2;
-    cout<<m[0][0].cost<<" () "<<m[0][0].parent;
-}
