--- L808_303/L808_303test.cpp
@@ -1,20 +1,0 @@
-#include "L808_303.hpp"
-#include <iostream>
-
-using namespace std;
-
-int string_compare(char *s, char *t);
-
-int main () {
-    const char* a="cr";
-    const char* b="lf";
-    for (size_t i=0; i<MAXLEN; ++i)
-    {
-        for (size_t j=0; j<MAXLEN; ++j)
-        {
-            m[i][j]=i+j;
-        }
-    }
-    cout<<string_compare(a, b);
-}
-
