+++ L808_303/L811_307.hpp
@@ -1,15 +1,8 @@
 #ifndef L811_307_h
 #define L811_307_h
 
-int match(char c, char d)
-{
-	if (c==d) return 0;
-	else return 1;
-}
+int match(char c, char d);
 
-int indel (char c)
-{
-	return 1;
-}
+int indel (char c);
 
 #endif //L811_307_h
