+++ L808_303/L812_307.hpp
@@ -1,10 +1,6 @@
 #ifndef L812_307_h
 #define L812_307_h
 
-goal_cell(char *s, char *t, int *i, int *j) //Изменён в L814_308.h
-{
-	*i = strlen(s)-1;
-	*j = strlen(t)-1;
-}
+goal_cell(char *s, char *t, int *i, int *j); //Изменён в L814_308.h
 
 #endif //L812_307_h
