+++ L811_307/L811_307.cpp
@@ -1,3 +1,8 @@
+#include "L811_307.hpp"
+#include <iostream>
+
+using namespace std;
+
 int match(char c, char d)
 {
 	if (c==d) return 0;
@@ -7,4 +12,10 @@ int match(char c, char d)
 int indel (char c)
 {
 	return 1;
+}
+
+int main() {
+    char* a="cr", b="lf";
+    cout<<match(a,b)<<match(a,a);
+    cout<<"\n"<<indel(a);
 }
