--- L812_307/L812_307test.cpp
@@ -1,13 +1,0 @@
-#include "L812_307.hpp"
-#include <iostream>
-
-using namespace std;
-
-void goal_cell(char *s, char *t, int *i, int *j);
-
-int main() {
-    char* a="cr", b="lf";
-    int* i=0, j=0;
-    goal_cell(a, b, &i, &j);
-    cout<<a[i]<<" "<<b[j];
-}
