--- L812_307/Makefile
@@ -1,15 +1,0 @@
-RM=del
-
-l812307o=L812_307.o L812_307test.o
-L812307.exe: $(l812307o)
-	$(CXX) -o L812307.exe $(l812307o)
-
-L812_307.o: L812_307.cpp
-	$(CXX) -c L812_307.cpp
-
-L812_307test.o: L812_307test.cpp L812_307.hpp
-	$(CXX) -c L812_307test.cpp
-
-.PHONY: clean
-clean:
-	$(RM) $(l812307o) L812307.exe
